+++ installer/phoenix/lib/openmaize_phx.ex
@@ -1,18 +1,20 @@
-defmodule Mix.Tasks.Openmaize.Gen.Phoenixauth do
+defmodule Mix.Tasks.Openmaize.Phx do
   use Mix.Task
 
-  import Mix.Openmaize
+  import Mix.Openmaize.Phx.Generator
 
   @moduledoc """
   Create modules for authorization and, optionally, email confirmation.
 
-  ## Options
+  ## Options and arguments
 
-  There is an optional argument which is used to set how the user
-  is indexed. The default value is "username". Calling this generator
-  with a different value will make the necessary changes to the user
-  file in the models directory, the user migrations file and the
-  session controller.
+  There is one argument:
+
+    * unique_id - "username", "email", etc.
+      * the default is "username"
+
+  This value will be used in the user file in the models directory,
+  the user migrations file and the session controller.
 
   There are two options:
 
@@ -26,16 +28,16 @@ defmodule Mix.Tasks.Openmaize.Gen.Phoenixauth do
   In the root directory of your project, run the following command (add `--confirm`
   if you want to create functions for email confirmation):
 
-      mix openmaize.gen.phoenixauth
+      mix openmaize.phx
 
   If you are using :email to identify (search for) your users, you need
   to add email to the command:
 
-      mix openmaize.gen.phoenixauth email
+      mix openmaize.phx email
 
   If you want to create files for an api, run the following command:
 
-      mix openmaize.gen.phoenixauth --api
+      mix openmaize.phx --api
 
   """
 
@@ -48,8 +50,7 @@ defmodule Mix.Tasks.Openmaize.Gen.Phoenixauth do
       uniq -> ":#{uniq}"
     end
 
-    srcdir = Path.join [Application.app_dir(:openmaize, "priv"),
-     "templates", "phoenix"]
+    srcdir = Path.expand("../templates", __DIR__)
 
     files = phx(opts[:api]) ++ case {opts[:api], opts[:confirm]} do
       {true, true} -> phx_api() ++ phx_confirm(true)
@@ -58,16 +59,32 @@ defmodule Mix.Tasks.Openmaize.Gen.Phoenixauth do
       _ -> phx_html()
     end
 
-    Mix.Openmaize.copy_files(srcdir, files, base: base_module(),
-      unique_id: unique_id, confirm: opts[:confirm], api: opts[:api])
+    copy_files(srcdir, files, base: base_module(), unique_id: unique_id,
+      confirm: opts[:confirm], api: opts[:api])
 
     Mix.shell.info """
 
-    Please check the generated files. You might need to uncomment certain
-    lines and / or change certain details, such as paths or user details.
+    We are almost ready!
 
-    Before you use Openmaize, you need to configure Openmaize.
-    See the documentation for Openmaize.Config for details.
+    Now edit the `mix.exs` file, adding `:openmaize` to the list of
+    `applications` and `{:openmaize, {"~> 2.7"}},` to the deps.
+    Then run `mix deps.get`.
+
+    You might also need to edit the database details in the `config`
+    files.
+
+    Then, to run the tests:
+
+        mix test
+
+    And to start the server:
+
+        mix phoenix.server
+
+    If you are not using "username" or "email" to index / identify
+    the user, and if any of the session_controller tests fail, you
+    might need to edit the `test/controllers/session_controller_test.exs`
+    and `test/support/test_helpers.ex` files.
     """
   end